requirements.txt
Django==2.1.5
cryptography
pymysql
djangorestframework